services/report_service/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --find-links=../shared_lib/dist requirements.in
#
--find-links ../shared_lib/dist

aiofiles==24.1.0
    # via cloud-sql-python-connector
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via cloud-sql-python-connector
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via starlette
attrs==25.3.0
    # via aiohttp
build==1.2.2.post1
    # via socialpulse-common
cachetools==5.5.2
    # via google-auth
certifi==2025.6.15
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
cloud-sql-python-connector[psycopg2]==1.18.2
    # via -r requirements.in
coverage[toml]==7.9.1
    # via
    #   -r requirements.in
    #   pytest-cov
cryptography==45.0.4
    # via cloud-sql-python-connector
db-dtypes==1.4.3
    # via -r requirements.in
dnspython==2.7.0
    # via cloud-sql-python-connector
fastapi==0.116.1
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
google-auth==2.40.3
    # via cloud-sql-python-connector
h11==0.16.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via yoyo-migrations
iniconfig==2.1.0
    # via pytest
multidict==6.5.0
    # via
    #   aiohttp
    #   yarl
numpy==2.3.1
    # via
    #   db-dtypes
    #   pandas
packaging==24.2
    # via
    #   build
    #   db-dtypes
    #   pytest
    #   socialpulse-common
pandas==2.3.0
    # via db-dtypes
parameterized==0.9.0
    # via -r requirements.in
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.1
    # via
    #   -r requirements.in
    #   socialpulse-common
psycopg2-binary==2.9.10
    # via -r requirements.in
pyarrow==20.0.0
    # via
    #   -r requirements.in
    #   db-dtypes
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.0
    # via socialpulse-common
pygments==2.19.2
    # via pytest
pyproject-hooks==1.2.0
    # via
    #   build
    #   socialpulse-common
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-cov
    #   socialpulse-common
pytest-cov==6.2.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via
    #   pydantic-settings
    #   socialpulse-common
pytz==2025.2
    # via pandas
requests==2.32.4
    # via cloud-sql-python-connector
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
socialpulse-common==0.1.0
    # via -r requirements.in
sqlparse==0.5.3
    # via yoyo-migrations
starlette==0.47.2
    # via fastapi
tabulate==0.9.0
    # via yoyo-migrations
typing-extensions==4.14.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via -r requirements.in
yarl==1.20.1
    # via aiohttp
yoyo-migrations==9.0.0
    # via -r requirements.in
zipp==3.23.0
    # via importlib-metadata
